backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
sentence-transformers
faiss-cpu
numpy
python-dotenv